Parameterized "City" as "{City}", "{The_City}", {the_City}"
To reduce ambiguity and assure consistent use across a document, it is
helpful to wrap defined terms in {}, e.g., {City}.  When referring to a
party as “the X” we parameterize as {the_City}.  Because some docs
refer by name, etc. It doesn’t work to say, for instance, “the Acme”.
Lots of flexibility and consistency.